mappers/pyproject.toml
[tool.poetry]
name = "mappers"
version = "0.1.0"
description = ""
authors = ["Alessandro Pacielli <alessandropacielli@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pandas = "^1.0.4"
paho-mqtt = "^1.5.0"
python-dotenv = "^0.13.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
